moving message identifier file to sidecar where it is used
